requirements.txt
dash==2.18.2
plotly==6.0.0
pandas==2.2.3
sqlalchemy==2.0.39
psycopg2-binary==2.9.10
python-dotenv==1.0.1